test(components): break up combobox test files
